requirements.txt
websockets>=12.0
requests>=2.31.0
python-dotenv>=1.0.0
pyinstaller>=6.0.0
